Function pallet_dkg_proposals::pallet::dispatchables::set_resource

pub fn set_resource<T: Config>(id: ResourceId, method: Vec<u8>)

Stores a method name on chain under an associated resource ID.

O(1) write

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::set_resource.