Module pallet_dkg_proposals::pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

The main components of this pallet are:

Pallet, which implements all of the dispatchable extrinsics of the pallet, among other public functions.

The subset of the functions that are dispatchable can be identified either in the dispatchables module or in the Call enum.

storage_types, which contains the list of all types that are representing a storage item. Otherwise, all storage items are listed among Type Definitions.

Config, which contains the configuration trait of this pallet.

Event and Error, which are listed among the Enums.

Re-exports

pub use __tt_error_token_1 as tt_error_token;

pub use __tt_default_parts_7 as tt_default_parts;

pub use __tt_extra_parts_7 as tt_extra_parts;

Modules

dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

storage_types

Auto-generated docs-only module listing all (public and private) defined storage types for this pallet.

Structs

DefaultForProposerThreshold

GenesisConfig

Can be used to configure the genesis state of this pallet.

Pallet

The Pallet struct, the main type that implements traits and standalone functions within the pallet.

Enums

Call

Contains a variant per dispatchable extrinsic that this pallet has.

Error

The Error enum of this pallet.

Event

The Event enum of this pallet

Traits

Config

Configuration trait of this pallet.

Functions

__type_value_for_default_for_proposer_threshold

Type Definitions

ChainNonces

All whitelisted chains and their respective transaction counts

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.

ProposalOf

ProposerCount

Number of proposers in set

ProposerThreshold

Number of votes required for a proposal to execute

Proposers

Tracks current proposer set

Resources

Utilized by the bridge software to map resource IDs to actual methods

VoterList

Votes

All known proposals. The key is the hash of the call and the deposit ID, to ensure it’s unique.

VotingKey

Tracks current ECDSA voting keys for each validator

VotingKeyTuple

VotingKeys

Storage type is [StorageValue] with value type BoundedVec < (T :: AccountId, VotingKey < T >), T :: MaxProposers >.