Trait pallet_dkg_proposals::pallet::Config

pub trait Config: Config + Config {
    type Event: From<Event<Self>> + IsType<<Self as Config>::Event>;
    type AdminOrigin: EnsureOrigin<Self::Origin>;
    type Proposal: Parameter + EncodeLike + EncodeAppend + Into<Vec<u8>> + AsRef<[u8]>;
    type NextSessionRotation: EstimateNextSessionRotation<Self::BlockNumber>;
    type DKGId: Member + Parameter + RuntimeAppPublic + MaybeSerializeDeserialize;
    type DKGAuthorityToMerkleLeaf: Convert<Self::DKGId, Vec<u8>>;
    type ChainIdentifier: Get<TypedChainId>;
    type ProposalLifetime: Get<Self::BlockNumber>;
    type ProposalHandler: ProposalHandlerTrait;
    type WeightInfo: WeightInfo;
}

The module configuration trait.

Associated Types

type Event: From<Event<Self>> + IsType<<Self as Config>::Event>

The overarching event type.

type AdminOrigin: EnsureOrigin<Self::Origin>

Origin used to administer the pallet

type Proposal: Parameter + EncodeLike + EncodeAppend + Into<Vec<u8>> + AsRef<[u8]>

Proposed transaction blob proposal

type NextSessionRotation: EstimateNextSessionRotation<Self::BlockNumber>

Estimate next session rotation

type DKGId: Member + Parameter + RuntimeAppPublic + MaybeSerializeDeserialize

Authority identifier type

type DKGAuthorityToMerkleLeaf: Convert<Self::DKGId, Vec<u8>>

Convert DKG AuthorityId to a form that would end up in the Merkle Tree.

For instance for ECDSA (secp256k1) we want to store uncompressed public keys (65 bytes) and later to Ethereum Addresses (160 bits) to simplify using them on Ethereum chain, but the rest of the Substrate codebase is storing them compressed (33 bytes) for efficiency reasons.

type ChainIdentifier: Get<TypedChainId>

The identifier for this chain. This must be unique and must not collide with existing IDs within a set of bridged chains.

type ProposalLifetime: Get<Self::BlockNumber>

type ProposalHandler: ProposalHandlerTrait

type WeightInfo: WeightInfo